requirements-lock.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements-lock.txt requirements-dev.txt
#
-e file:///Users/sami/PycharmProjects/parquet-converter
    # via file:///users/sami/pycharmprojects/parquet-converter
black==25.1.0
    # via
    #   -r requirements-dev.txt
    #   parquet-converter
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via
    #   black
    #   pip-tools
coverage[toml]==7.8.0
    # via pytest-cov
cramjam==2.10.0
    # via fastparquet
fastparquet==2024.11.0
    # via parquet-converter
flake8==7.2.0
    # via
    #   -r requirements-dev.txt
    #   parquet-converter
fsspec==2025.3.2
    # via fastparquet
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via
    #   -r requirements-dev.txt
    #   parquet-converter
mccabe==0.7.0
    # via flake8
mypy==1.15.0
    # via
    #   -r requirements-dev.txt
    #   parquet-converter
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
numpy==2.2.5
    # via
    #   fastparquet
    #   pandas
packaging==25.0
    # via
    #   black
    #   build
    #   fastparquet
    #   pytest
pandas==2.2.3
    # via
    #   fastparquet
    #   parquet-converter
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via
    #   -r requirements-dev.txt
    #   parquet-converter
platformdirs==4.3.8
    # via black
pluggy==1.5.0
    # via pytest
pyarrow==20.0.0
    # via parquet-converter
pycodestyle==2.13.0
    # via flake8
pyflakes==3.3.2
    # via flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.5
    # via
    #   -r requirements-dev.txt
    #   parquet-converter
    #   pytest-cov
pytest-cov==6.1.1
    # via
    #   -r requirements-dev.txt
    #   parquet-converter
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via parquet-converter
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via parquet-converter
six==1.17.0
    # via python-dateutil
tqdm==4.67.1
    # via parquet-converter
typing-extensions==4.13.2
    # via mypy
tzdata==2025.2
    # via pandas
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
